HU5 Completa, JWT

+++ src/main/java/com/pragma/powerup/domain/model/UserModel.java
@@ -20,5 +20,5 @@ public class UserModel {
     private LocalDate birthDate;
     private String email;
     private String password;
-    private int idRole;
+    private Long idRole;
 }
